WAITP-1214 tweaks to PR
