+++ packages/tupaia-web/src/constants/index.ts
@@ -5,3 +5,4 @@
 export * from './colors';
 export * from './constants';
 export * from './mapbox';
+export * from './url';
