+++ packages/tupaia-web/src/layout/Layout.tsx
@@ -8,7 +8,7 @@ import styled from 'styled-components';
 import { Outlet } from 'react-router';
 import { EnvBanner } from '@tupaia/ui-components';
 import { TopBar } from './TopBar';
-import { MOBILE_THRESHOLD } from '../constants';
+import { MOBILE_BREAKPOINT } from '../constants';
 
 /**
  * This is the layout for the entire app, which contains the top bar and the main content. This is used to wrap the entire app content
@@ -26,7 +26,7 @@ const Container = styled.div`
   svg.recharts-surface {
     overflow: visible;
   }
-  @media screen and (min-width: ${MOBILE_THRESHOLD}) {
+  @media screen and (min-width: ${MOBILE_BREAKPOINT}) {
     overflow-y: hidden;
   }
 `;
