+++ packages/tupaia-web/src/layout/Map/MapOverlaySelector/MobileMapOverlaySelector.tsx
@@ -6,12 +6,12 @@ import React, { useState } from 'react';
 import styled from 'styled-components';
 import { ArrowBackIos, ArrowForwardIos } from '@material-ui/icons';
 import { Button } from '@tupaia/ui-components';
-import { MOBILE_THRESHOLD } from '../../../constants';
+import { MOBILE_BREAKPOINT } from '../../../constants';
 
 // Placeholder for MapOverlaySelector component
 const Wrapper = styled.div`
   width: 100%;
-  @media screen and (min-width: ${MOBILE_THRESHOLD}) {
+  @media screen and (min-width: ${MOBILE_BREAKPOINT}) {
     display: none;
   }
 `;
@@ -58,7 +58,7 @@ const OverlayMenu = styled.div<{
   }
 `;
 
-export const ExpandableMapOverlaySelector = () => {
+export const MobileMapOverlaySelector = () => {
   const [expanded, setExpanded] = useState(false);
   const toggleExpanded = () => setExpanded(!expanded);
   return (
