+++ packages/tupaia-web/src/layout/TabLayout/index.ts
@@ -3,4 +3,4 @@
  * Copyright (c) 2017 - 2023 Beyond Essential Systems Pty Ltd
  */
 
-export { TabLayout } from './TabLayout';
+export { MobileTabLayout } from './MobileTabLayout';
