+++ packages/tupaia-web/src/pages/LandingPage/LandingPageFooter.tsx
@@ -7,7 +7,7 @@ import React from 'react';
 import styled from 'styled-components';
 import { Link, Typography, List } from '@material-ui/core';
 import { SingleLandingPage } from '../../types';
-import { MOBILE_THRESHOLD } from '../../constants';
+import { MOBILE_BREAKPOINT } from '../../constants';
 
 const Footer = styled.footer`
   margin-top: auto;
@@ -19,7 +19,7 @@ const Footer = styled.footer`
 const FooterContentWrapper = styled.div`
   display: flex;
   flex-direction: column;
-  @media screen and (min-width: ${MOBILE_THRESHOLD}) {
+  @media screen and (min-width: ${MOBILE_BREAKPOINT}) {
     flex-direction: row;
   }
 `;
@@ -48,7 +48,7 @@ const FooterContentContainer = styled.div`
   & + & {
     margin-top: 1em;
   }
-  @media screen and (min-width: ${MOBILE_THRESHOLD}) {
+  @media screen and (min-width: ${MOBILE_BREAKPOINT}) {
     width: 30%;
     & + & {
       margin-left: 4em;
